// src/app/layouts/sante/covid/services/models/covid.ts
export interface Covid {
    id?: string;
    PhotoUrl?: string;
    NbreTotalContamine?: number;
    NbreContaminerFemme?: number;
    NbreTotalGueris?: number;
    NbreGuerisFemme?: number;
    NbreTotalDeces?: number;
    NbreDecesFemme?: number;
    EtatFinancementStructureSanitaire?: string;
    AccesbiliteStructureSantairePopulation?: string;
    RapportGeneraleCovid?: string;
    Province?: string;
    CommunaleUrbain?: string;
    CheferieSecteur?: string;
    Created?: any;
    agentId?: string;
}